feat: FavoriteStations를 별도의 클래스로 분리
